Uso de custom errors, corrección de sesión en CRUD tareas

+++ src/modules/task/task.service.ts
@@ -12,14 +12,18 @@ export class TaskService {
 
   private tasks: any[] = [];
 
-  async getTasks() {
-    const tasks = await this.prisma.task.findMany();
+  async getTasks(user_id: number): Promise<Task[]> {
+    const tasks = await this.prisma.task.findMany({
+      where: {
+        user_id
+      }
+    });
     return tasks;
   }
 
-  async getTaskById(id: number): Promise<Task> {
+  async getTaskById(id: number, user_id: number): Promise<Task> {
     const task = await this.prisma.task.findUnique({
-      where: { id },
+      where: { id, user_id },
     });
 
     if (task == undefined) {
@@ -29,22 +33,26 @@ export class TaskService {
     return task;
   }
 
-  async insertTask(task: CreateTaskDto): Promise<Task> {
-    //Agregar el query
-    const newTask = await this.prisma.task.create({ data: task });
+  async insertTask(task: CreateTaskDto, user_id: number): Promise<Task> {
+    const newTask = await this.prisma.task.create({
+      data: {
+        ...task,
+        user_id
+      }
+    });
 
     return newTask;
   }
 
-  async updateTask(id: number, taskUpdate: UpdateTaskDto): Promise<Task> {
-    const task = await this.prisma.task.update({ where: { id }, data: taskUpdate });
+  async updateTask(id: number, user_id: number, taskUpdate: UpdateTaskDto): Promise<Task> {
+    const task = await this.prisma.task.update({ where: { id, user_id }, data: taskUpdate });
 
     return task;
   }
 
-  async deleteTask(id: number): Promise<boolean> {
-  const deletedTask = await this.prisma.task.delete({ where: { id } });
+  async deleteTask(id: number, user_id:number): Promise<boolean> {
+    const deletedTask = await this.prisma.task.delete({ where: { id, user_id } });
 
-  return deletedTask ? true : false;
+    return deletedTask ? true : false;
   }
 }
